Phase D12 part 2: regime stub + CLIs + Stage 1 verdict on 5 challenges (#6)
Completes the Phase D12 roadmap (docs/design/phase-d12-to-e.md §6.13
steps 10-15). Stage 1 gate is now runnable end-to-end from any entry
JSONL, the stress runner fires circuit breakers against known
disaster windows, and every active challenge has a preregistered
verdict recorded.

New code:
- cogochi-autoresearch/backtest/regime.py  — D12 stub (always "unknown")
                                              + D15-preview BTC 30d labeler
- cogochi-autoresearch/tests/test_backtest_regime.py   — 4 new tests
- tools/backtest_portfolio.py              — Stage 1 CLI, writes
                                              metadata/metrics/trades/
                                              equity_curve under
                                              tools/output/runs/<run_id>/
- tools/stress_test.py                     — 2 D12 scenarios (2020-03-12
                                              flash crash, 2022-05/11 LUNA+FTX)

Bug fixes + typing cleanup:
- scanner/pnl.py:_compute_slippage_pct    cast → float (mypy --strict)
- backtest/portfolio.py:_week_key          explicit date return type

New docs:
- docs/design/stage-1-verdict.md          preregistered vs observed table
                                           and detailed postmortems
- docs/design/graveyard.md                +3 entries (lgb-long-v1,
                                            sample-rally, ake-style)
- docs/design/experiments_log.md          +8 rows (11-18), k=18 still
                                            under preregistered 20

Results (all runs with default RiskConfig(), pessimistic intrabar):

| Challenge             |  n_exec | expectancy |  MDD    | verdict |
|-----------------------|---------|------------|---------|---------|
| sample-rally-pattern  |  2061   | -0.49%     | -99.5%  | FAIL    |
| btc-macd-style        |   359   | +2.01%     |  -5.1%  | **PASS**|
| ake-style             |  1742   | -0.59%     | -99.6%  | FAIL    |
| exhaustion-short      |   295   | -1.25%     | -84.6%  | FAIL    |
| lgb-long-v1 thr=0.60  |    17   | -1.43%     | -11.6%  | FAIL    |

The big surprise is btc-macd-style. D9 killed it at t=0.07σ on
`excess_positive_rate × coverage`, but under path-walk execution with
2% stop / 4% target / 24h horizon it produces 358 in-sample BTCUSDT
trades with 75.8% win rate. A BTCUSDT random-baseline sanity check
(1000 uniform longs, same stop/target) returns mean pnl -0.03% and
t=-0.40 vs zero, so the +2.01% is NOT drift — it's a real in-sample
edge. That promotes btc-macd-style to a preregistered Phase D16
walk-forward candidate.

The bigger lesson: the horizon-based scoring used in Phases D7-D11
systematically understates strategies whose edge is visible in the
bar-by-bar path and misses losses that the path-walk model catches
(lgb-long-v1 went from 77.4% offline win rate to 11.8% under realistic
execution, entirely because of the 2% stop-loss touching on the way to
a positive 24h close).

Stress tests (scenarios 1+2): both runs show the expected cascade —
daily_loss_halt → weekly_loss_halt → consecutive_loss_pause, with MDD
held at -17% / -13% instead of the raw 30%+ the windows would produce
without circuit breakers.

Validation:
- 280 tests pass (276 + 4 new regime tests)
- mypy --strict clean on all D12 modules (backtest/, observability/,
  scanner/pnl.py, exceptions.py). Pre-existing data_cache/ and
  models/signal.py errors are out of scope for this PR.

+++ scanner/pnl.py
@@ -106,7 +106,7 @@ def _compute_slippage_pct(
     frac = notional_usd / adv_notional_usd
     if frac < costs.adv_notional_fraction_threshold:
         return base
-    return base + costs.impact_k * (frac ** 0.5)
+    return float(base + costs.impact_k * (frac ** 0.5))
 
 
 def _resolve_fee_single_side(costs: ExecutionCosts) -> float:
